Added lots of stuff

+++ args.cc
@@ -10,9 +10,9 @@ using namespace std;
 arguments::arguments (const int argc,
                       const char* const* const argv)
 	: image_width (0),
-	  view_width (3.0f),
-	  hcenter (-0.5f),
-	  vcenter (0.0f),
+	  view_width (3.0),
+	  hcenter (-0.5),
+	  vcenter (0.0),
 	  hsample (1),
 	  vsample (1)
 {
@@ -23,21 +23,21 @@ arguments::arguments (const int argc,
 
 	if (argc < 4)
 		return;
-	view_width = stof (argv [3]);
+	view_width = stod (argv [3]);
 
 	if (argc < 5)
 		return;
-	hcenter = stof (argv [4]);
+	hcenter = stod (argv [4]);
 
 	if (argc < 6)
 		return;
-	vcenter = stof (argv [5]);
+	vcenter = stod (argv [5]);
 
 	if (argc < 7)
 		return;
-	hsample = stof (argv [6]);
+	hsample = stod (argv [6]);
 
 	if (argc < 8)
 		return;
-	vsample = stof (argv [7]);
+	vsample = stod (argv [7]);
 }
